pyproject.toml
[project]
name = "erlesen"
version = "0.1.0"
description = ""
authors = [
    {name = "Lars Klöser",email = "kloeser@fh-aachen.de"}
]
readme = "README.md"
requires-python = ">=3.12,<3.13"
dependencies = [
    "spacy (>=3.8.4,<4.0.0)",
    "openai (>=1.60.2,<2.0.0)",
    "python-dotenv (>=1.0.1,<2.0.0)",
    "torch (>=2.6.0,<3.0.0)",
    "transformers (>=4.48.1,<5.0.0)",
    "gradio (>=5.13.2,<6.0.0)",
    "textstat (>=0.7.4,<0.8.0)",
    "evaluate (>=0.4.3,<0.5.0)",
    "peft (>=0.14.0,<0.15.0)",
    "scikit-learn (>=1.6.1,<2.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
